refactor(conftest): Move to libtmux/conftest.py
